refactored background.js to split into separate modules based on functionalities
